+++ components/mypage/posts/MoabangPostList.tsx
@@ -3,10 +3,7 @@
 import { useEffect, useRef, useCallback } from 'react';
 import Link from 'next/link';
 import Pagination from '@/components/common/pagination/Pagination';
-import {
-  useMyMoabangPostsQuery,
-  useMyMoabangPostsInfiniteQuery,
-} from '@/hooks/queries/user/useMyPosts';
+import { useMyMoabangPostsQuery, useMyMoabangPostsInfiniteQuery } from '@/hooks/queries/user';
 import { EmptyState } from '@/components/common/empty-state/EmptyState';
 import { useIsDesktop } from '@/hooks/useIsMobile';
 import { MyMoabangPost } from './post.type';
